Add utility functions to convert SortOrder and OrderingScheme
