+++ presto-main/src/main/java/com/facebook/presto/sql/planner/PlannerUtils.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.sql.planner;
+
+import com.facebook.presto.spi.block.SortOrder;
+import com.facebook.presto.sql.tree.SortItem;
+
+public class PlannerUtils
+{
+    private PlannerUtils()
+    {}
+
+    public static SortOrder toSortOrder(SortItem sortItem)
+    {
+        if (sortItem.getOrdering() == SortItem.Ordering.ASCENDING) {
+            if (sortItem.getNullOrdering() == SortItem.NullOrdering.FIRST) {
+                return SortOrder.ASC_NULLS_FIRST;
+            }
+            return SortOrder.ASC_NULLS_LAST;
+        }
+        if (sortItem.getNullOrdering() == SortItem.NullOrdering.FIRST) {
+            return SortOrder.DESC_NULLS_FIRST;
+        }
+        return SortOrder.DESC_NULLS_LAST;
+    }
+}
